Create Day2CompoundEvent.py
There are  urns labeled , , and . 

Urn  X contains  red balls and  black balls.
Urn  Y contains  red balls and  black balls.
Urn  Z contains  red balls and  black balls. 

One ball is drawn from each of the  urns. What is the probability that, of the  balls drawn,  2 are red and 1  is black?